requirements.txt
appdirs==1.4.3
attrs==19.3.0
-e git+https://github.com/suhrawardi/behave.git@a5ce707c4c49c51a403b67d916c27c416ebcbf5e#egg=behave
behave2cucumber==1.0.3
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cucumber-tag-expressions==2.0.4
defusedxml==0.6.0
distlib==0.3.0
filelock==3.0.12
idna==2.9
isodate==0.6.0
lxml==4.5.0
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
-e git+https://github.com/suhrawardi/python-odata.git@4f063673710d8e480f8825c170b7aa0db631e0cf#egg=odata
parse==1.15.0
parse-type==0.5.2
pipenv==2018.11.26
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
PyHamcrest==2.0.2
pylama==7.7.1
pync==2.0.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.14.0
snowballstemmer==2.0.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.13
virtualenv-clone==0.5.3
zeep==3.4.0
